Fix: Error comment
Comment: "Is there any point in adding an id? On the other hand whenever rending a list in React, it is necessary to add a key prop, so that react can keep track of which is which between re-renders. If you open the chrome debugging inspector you will see the following error:"